+++ config/index.js
@@ -3,7 +3,7 @@
 // see http://vuejs-templates.github.io/webpack for documentation.
 
 const path = require('path');
-const backEndAddr = '';
+const backEndAddr = 'http://frontend.8slan.com/';
 
 module.exports = {
   dev: {
