+++ src/components/projects/workflow/productCustomTaskDetail/jobBuildDetail.vue
@@ -1,15 +1,15 @@
 <template>
-  <div class="common-task">
+  <div class="job-build-detail">
     <header class="mg-b8">
       <el-col :span="6">
-        <span class="type">通用任务</span>
-        <span>{{commonInfo.name}}</span>
+        <span class="type">构建</span>
+        <span>{{jobInfo.name}}</span>
       </el-col>
       <el-col :span="2">
-        <a :class="buildOverallColor" href="#buildv4-log">{{commonInfo.status?buildOverallStatusZh:"未运行"}}</a>
+        <a :class="buildOverallColor" href="#buildv4-log">{{jobInfo.status?buildOverallStatusZh:"未运行"}}</a>
       </el-col>
       <el-col :span="2">
-        <span>{{commonInfo.interval}}</span>
+        <span>{{jobInfo.interval}}</span>
       </el-col>
       <el-col :span="1" class="close">
         <span @click="$emit('showFooter',false)">
@@ -20,9 +20,9 @@
     <main>
       <section>
         <div class="error-wrapper">
-          <el-alert v-if="commonInfo.error" title="错误信息" :description="commonInfo.error" type="error" close-text="知道了"></el-alert>
+          <el-alert v-if="jobInfo.error" title="错误信息" :description="jobInfo.error" type="error" close-text="知道了"></el-alert>
         </div>
-        <el-row class="text item mg-t8" :gutter="0" v-for="(build,index) in commonInfo.spec.repos" :key="index">
+        <el-row class="text item mg-t8" :gutter="0" v-for="(build,index) in jobInfo.spec.repos" :key="index">
           <el-col :span="4">
             <div class="grid-content item-title">代码库({{build.source}})</div>
           </el-col>
@@ -36,9 +36,42 @@
             <RepoJump :build="build" showIcon />
           </el-col>
         </el-row>
+        <el-row :gutter="0" class="mg-t8">
+          <el-col :span="4">
+            <div class="item-title">服务名称</div>
+          </el-col>
+          <el-col :span="8">
+            <span>{{jobInfo.spec.service_name}}({{jobInfo.spec.service_module}})</span>
+          </el-col>
+          <el-col :span="4">
+            <div class="item-title">
+              镜像名称
+              <el-tooltip effect="dark" placement="top">
+                <div slot="content">
+                  构建镜像标签生成规则 ：
+                  <br />选择 Tag 进行构建 ： 构建时间戳 -
+                  Tag
+                  <br />只选择分支进行构建：构建时间戳
+                  - 任务 ID - 分支名称
+                  <br />选择分支和 PR 进行构建：构建时间戳 - 任务 ID - 分支名称 - PR ID
+                  <br />只选择 PR
+                  进行构建：构建时间戳 - 任务 ID - PR ID
+                </div>
+                <span>
+                  <i class="el-icon-question"></i>
+                </span>
+              </el-tooltip>
+            </div>
+          </el-col>
+          <el-col :span="8">
+            <el-tooltip effect="dark" :content="jobInfo.spec.image" placement="top">
+              <span>{{jobInfo.spec.image.split('/').pop()}}</span>
+            </el-tooltip>
+          </el-col>
+        </el-row>
       </section>
       <section class="log-content mg-t8">
-        <XtermLog :id="commonInfo.name" @mouseleave.native="leaveLog" :logs="buildv4AnyLog" :from="commonInfo.name" />
+        <XtermLog :id="jobInfo.name" @mouseleave.native="leaveLog" :logs="buildv4AnyLog" :from="jobInfo.name" />
       </section>
     </main>
   </div>
@@ -60,7 +93,7 @@ export default {
     }
   },
   props: {
-    commonInfo: {
+    jobInfo: {
       type: Object,
       default: () => ({})
     },
@@ -75,6 +108,10 @@ export default {
     workflowName: {
       type: String,
       default: ''
+    },
+    isShowConsoleFooter: {
+      type: Boolean,
+      default: false
     }
   },
   mixins: [mixin],
@@ -83,13 +120,13 @@ export default {
   },
   computed: {
     buildIsRunning () {
-      return this.commonInfo && this.commonInfo.status === 'running'
+      return this.jobInfo && this.jobInfo.status === 'running'
     },
     buildIsDone () {
-      return this.isSubTaskDone(this.commonInfo)
+      return this.isSubTaskDone(this.jobInfo)
     },
     buildOverallStatus () {
-      return this.$utils.calcOverallBuildStatus(this.commonInfo, {})
+      return this.$utils.calcOverallBuildStatus(this.jobInfo, {})
     },
     buildOverallStatusZh () {
       return this.$translate.translateTaskStatus(this.buildOverallStatus)
@@ -100,16 +137,16 @@ export default {
   },
   methods: {
     leaveLog () {
-      const el = document.querySelector('.common-task').style
+      const el = document.querySelector('.job-build-detail').style
       el.overflow = 'auto'
     },
     openBuildLog (buildType) {
       this.buildv4AnyLog = []
-      const url = `/api/aslan/logs/sse/v4/workflow/${this.workflowName}/${this.taskId}/${this.commonInfo.name}/999999?projectName=${this.projectName}`
+      const url = `/api/aslan/logs/sse/v4/workflow/${this.workflowName}/${this.taskId}/${this.jobInfo.name}/999999?projectName=${this.projectName}`
       if (typeof window.msgServer === 'undefined') {
         window.msgServer = {}
         window.msgServer[
-          `${this.commonInfo.spec.service_module}_${this.commonInfo.spec.service_name}`
+          `${this.jobInfo.spec.service_module}_${this.jobInfo.spec.service_name}`
         ] = {}
       }
       this[`${buildType}IntervalHandle`] = setInterval(() => {
@@ -123,7 +160,7 @@ export default {
         .then(sse => {
           // Store SSE object at a higher scope
           window.msgServer[
-            `${this.commonInfo.spec.service_module}_${this.commonInfo.spec.service_name}`
+            `${this.jobInfo.spec.service_module}_${this.jobInfo.spec.service_name}`
           ] = sse
           sse.onError(e => {
             console.error('lost connection; giving up!', e)
@@ -147,7 +184,7 @@ export default {
       return getHistoryLogsAPI(
         this.workflowName,
         this.taskId,
-        this.commonInfo.name,
+        this.jobInfo.name,
         this.projectName
       ).then(response => {
         this.buildv4AnyLog = response.split('\n').map(element => {
@@ -168,10 +205,13 @@ export default {
           ? Math.round(new Date().getTime() / 1000)
           : detail.end_time) - detail.start_time
       detail.interval = this.$utils.timeFormat(detail.intervalSec)
+    },
+    closeConsole () {
+      this.isShowConsoleFooter = false
     }
   },
   watch: {
-    commonInfo: {
+    jobInfo: {
       handler (val, oldVal) {
         if (val) {
           if (oldVal && val.name !== oldVal.name) {
@@ -206,7 +246,7 @@ export default {
 }
 </script>
 <style lang="less" scoped>
-.common-task {
+.job-build-detail {
   position: relative;
   height: 100%;
   font-size: 14px;
